package structure changed and 'blog' removed from the package name
