+++ src/main/groovy/de/triology/testdata/builder/EntityBuilderScript.groovy
@@ -21,19 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import de.triology.testdata.loader.testentities.AnotherTestEntity
-import de.triology.testdata.loader.testentities.BasicTestEntity
-import de.triology.testdata.loader.testentities.TestEntityWithToOneRelationship
 
-create TestEntityWithToOneRelationship, 'entityWithToOneRelationship', {
-    referencedEntity = create BasicTestEntity, 'referencedInstance', {
-        stringProperty = 'string in referenced entity'
-        integerProperty = 222
-    }
-}
+package de.triology.testdata.builder
 
-create AnotherTestEntity, 'entityOfAnotherClass', {}
+abstract class EntityBuilderScript extends Script {
 
-create TestEntityWithToOneRelationship, 'anotherEntityWithToOneRelationship', {
-    referencedEntity = referencedInstance
+    @Delegate
+    @Lazy
+    EntityBuilderDsl dsl = this.binding.dsl
 }
