--- src/test/java/de/triology/blog/testdataloader/EntityPersisterTest.java
@@ -1,43 +1,0 @@
-/**
- * MIT License
- *
- * Copyright (c) 2016 TRIOLOGY GmbH
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package de.triology.blog.testdataloader;
-
-import org.junit.Test;
-
-import javax.persistence.EntityManager;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-
-public class EntityPersisterTest {
-
-    @Test
-    public void persistsCreatedEntities() throws Exception {
-        EntityManager entityManagerMock = mock(EntityManager.class);
-        EntityPersister entityPersister = new EntityPersister(entityManagerMock);
-        Object entity = new Object();
-        entityPersister.onEntityCreated("entity1", entity);
-        verify(entityManagerMock).persist(entity);
-    }
-}
