+++ src/test/resources/tests/testEntityDefinitions.groovy
@@ -21,12 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import de.triology.blog.testdataloader.testentities.AnotherTestEntity
-import de.triology.blog.testdataloader.testentities.BasicTestEntity
-import de.triology.blog.testdataloader.testentities.TestEntityWithToManyRelationship
-import de.triology.blog.testdataloader.testentities.TestEntityWithToOneRelationship
-
 import java.text.SimpleDateFormat
+
+import de.triology.testdata.loader.testentities.AnotherTestEntity
+import de.triology.testdata.loader.testentities.BasicTestEntity
+import de.triology.testdata.loader.testentities.TestEntityWithToManyRelationship
+import de.triology.testdata.loader.testentities.TestEntityWithToOneRelationship
 
 create BasicTestEntity, 'basicEntity', {
     stringProperty = 'a string value'
